# src/main/java/com/josue/kodeur/xtremanalyse/application/entities/lieux/Prefecture.java
package com.josue.kodeur.xtremanalyse.application.entities.lieux;



import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

import javax.persistence.*;
import java.io.Serializable;
import java.time.LocalDateTime;

@AllArgsConstructor
@NoArgsConstructor
@Getter @Setter
@Entity @Table(name = "prefectures")
public class Prefecture implements Serializable {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    @Column(name = "id", nullable = false)
    private Long id;

    @Column(nullable = false, length = 40)
    private String nom;

    @ManyToOne
    @JoinColumn(name = "region_id")
    private Region region;

    @Column(name = "created_at", nullable = false)
    private LocalDateTime createdAt;

    @Column(name = "update_at", nullable = false)
    private LocalDateTime updateAt;
}